converter refactor
