+++ src/main/java/com/rbkmoney/analytics/converter/CostToAmountResponseConverter.java
@@ -1,11 +1,8 @@
 package com.rbkmoney.analytics.converter;
 
 import com.rbkmoney.analytics.dao.model.NumberModel;
-import com.rbkmoney.analytics.dao.model.ShopAmountModel;
 import com.rbkmoney.damsel.analytics.AmountResponse;
 import com.rbkmoney.damsel.analytics.CurrencyGroupedAmount;
-import com.rbkmoney.damsel.analytics.ShopAmountResponse;
-import com.rbkmoney.damsel.analytics.ShopGroupedAmount;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,17 +21,6 @@ public class CostToAmountResponseConverter {
 
         return new AmountResponse()
                 .setGroupsAmount(collect);
-    }
-
-    public ShopAmountResponse convertShop(List<ShopAmountModel> shopAmountModels) {
-        List<ShopGroupedAmount> collect = shopAmountModels.stream()
-                .map(cost -> new ShopGroupedAmount()
-                        .setAmount(cost.getAmount())
-                        .setShopId(cost.getShopId())
-                        .setCurrency(cost.getCurrency())
-                ).collect(toList());
-
-        return new ShopAmountResponse().setGroupsAmount(collect);
     }
 
 }
